fix(po-rep): properly calculate cumulative USD history values
- Updated SQL queries to change column aliases for clarity in `getFilecoinPaymentsForDealsHistory.sql` and `getPoRepDealsValueHistory.sql`.
- Removed cumulative amount calculations from SQL queries in `getFilecoinPaymentsForDealsHistory.sql` and `getPoRepDealsValueHistory.sql`.
- Refactored `PoRepController` to streamline data retrieval methods, replacing manual mapping with direct service calls.
- Consolidated and organized types in `types.po-rep.ts`, removing redundant interfaces and ensuring consistency across parameters.
- Enhanced type safety and clarity in the `PoRepService` methods for fetching deals and compliance history.
